pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feedshift"
version = "0.6.0"
description = "Intelligent recommendation engine that transforms noisy social feeds into personalized, semantically relevant, and high-quality content streams"
authors = [
    {name = "Amit Joshi", email = "amitjoshi937@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["recommendation", "ai", "social-media", "content-filtering", "feedshift", "personalization"]


classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=2.2.3",
    "numpy>=2.0.1",
    "torch>=2.6.0",
    "scikit-learn>=1.6.1",
    "detoxify>=0.5.2",
    "reflex==0.7.14",
    "streamlit>=1.45.1",
    "praw>=7.8.1",
    "python-dotenv>=1.1.0",
    "transformers==4.51.3",
    "sentence-transformers==4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/amitjoshi9627/FeedShift"
Repository = "https://github.com/amitjoshi9627/FeedShift"
Issues = "https://github.com/amitjoshi9627/FeedShift/issues"
Changelog = "https://github.com/amitjoshi9627/FeedShift/releases"


[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "dashboard*", "assets*"]
exclude = ["tests*", "examples*"]


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/conftest.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | examples
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.flake8]
max-line-length = 120
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
    "E501",  # line too long
    "F401",  # imported but unused (we'll fix these manually)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
    ".tox",
    ".mypy_cache",
    "examples",
]
